FormatOps: move getApplyArgs from TreeOps
We are about to start using some functionality currently in FormatOps.
